services/api/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
kafka-python==2.0.2
pydantic==2.11.1
